+++ packages/aw-signer/src/lib/utils/pkp-tool-registry/index.ts
@@ -1,3 +1,3 @@
 export * from './pkp-tool-registry';
-export * from './get-registered-tool-and-delegatees';
 export * from './get-permitted-tools-for-delegatee';
+export * from './get-registered-tool-and-delegatees';
