+++ packages/aw-tool-jupiter-swap/src/lib/lit-actions/utils/solana-connection.ts
@@ -1,0 +1,8 @@
+import { Connection } from "@solana/web3.js";
+
+export function createSolanaConnection(endpoint = 'https://api.mainnet-beta.solana.com'): Connection {
+    return new Connection(endpoint, {
+      commitment: 'confirmed',
+      confirmTransactionInitialTimeout: 10000
+    });
+  }
