+++ packages/aw-tool-sign-eddsa/src/lib/lit-actions/utils/index.ts
@@ -1,0 +1,2 @@
+export * from './solana-keypair';
+export * from './sign-message';
